move redux files into redux folder
