+++ client/src/redux/store.ts
@@ -1,0 +1,10 @@
+import { configureStore } from "@reduxjs/toolkit";
+import userReducer from "./user";
+
+const store = configureStore({
+  reducer: {
+    user: userReducer,
+  },
+});
+
+export default store;
