--- client/src/store.ts
@@ -1,26 +1,0 @@
-import { configureStore, createSlice } from "@reduxjs/toolkit";
-
-const initialState = {
-  user: null,
-};
-
-const userSlice = createSlice({
-  name: "user",
-  initialState,
-  reducers: {
-    setUser(state, action) {
-      state.user = action.payload;
-    },
-    clearUser(state) {
-      state.user = null;
-    },
-  },
-});
-
-const store = configureStore({
-  reducer: {
-    user: userSlice.reducer,
-  },
-});
-
-export default store;
